refactor: organise frontend components folders
