+++ frontend/src/pages/DeletedProductPage.jsx
@@ -1,4 +1,4 @@
-import ProductDeletedShowInfo from "../components/ProductDeletedShowInfo";
+import ProductDeletedShowInfo from "../components/product/ProductDeletedShowInfo";
 
 export default function DeletedProductPage() {
   return <ProductDeletedShowInfo />;
