+++ frontend/src/pages/DeletedUserPage.jsx
@@ -1,4 +1,4 @@
-import UserDeletedShowInfo from "../components/UserDeletedShowInfo";
+import UserDeletedShowInfo from "../components/user/UserDeletedShowInfo";
 
 export default function DeletedUserPage() {
   return <UserDeletedShowInfo />;
