+++ frontend/src/pages/UserPage.jsx
@@ -1,4 +1,4 @@
-import UserShowInfo from "../components/UserShowInfo";
+import UserShowInfo from "../components/user/UserShowInfo";
 
 export default function UserPage() {
   return <UserShowInfo />;
