+++ frontend/src/pages/UsersPage.jsx
@@ -1,4 +1,4 @@
-import UsersValid from "../components/UsersValid";
+import UsersValid from "../components/user/UsersValid";
 
 export default function UsersPage() {
   return <UsersValid />;
